web/refactor: functions to misc & nav folder
