+++ web/app/components/Navbar/Logo.tsx
